moved mxl parser to new file
